Move CLI script tests into SiteScanTest
