+++ tests/php/src/Helpers/ValidationRequestMocking.php
@@ -1,0 +1,63 @@
+<?php
+
+namespace AmpProject\AmpWP\Tests\Helpers;
+
+use AMP_Validated_URL_Post_Type;
+use WP_Query;
+
+/**
+ * Class ValidationRequestMocking
+ *
+ * Helpers for validation tests tests.
+ */
+final class ValidationRequestMocking {
+
+	/**
+	 * Gets all of the validated URLs.
+	 *
+	 * @return string[] $urls The validated URLs.
+	 */
+	public static function get_validated_urls() {
+		$query = new WP_Query(
+			[
+				'post_type'      => AMP_Validated_URL_Post_Type::POST_TYPE_SLUG,
+				'posts_per_page' => 100,
+				'fields'         => 'ids',
+			]
+		);
+
+		return array_map(
+			static function( $post ) {
+				return remove_query_arg( 'amp', AMP_Validated_URL_Post_Type::get_url_from_post( $post ) );
+			},
+			$query->posts
+		);
+	}
+
+	/**
+	 * Construct a WP HTTP response for a validation request.
+	 *
+	 * @return array The response.
+	 */
+	public static function get_validate_response() {
+		$mock_validation = [
+			'results' => [
+				[
+					'error'     => [
+						'code' => 'foo',
+					],
+					'sanitized' => false,
+				],
+			],
+			'url'     => home_url( '/' ),
+		];
+
+		return [
+			'body'     => wp_json_encode( $mock_validation ),
+			'response' => [
+				'code'    => 200,
+				'message' => 'ok',
+			],
+		];
+	}
+}
